pyproject.toml
[project]
name = "cot-faithfulness-model-diffing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.0.0",
    "dotenv>=0.9.9",
    "huggingface-hub>=0.34.4",
    "ipykernel>=6.30.1",
    "mlcroissant>=1.0.22",
    "nnsight>=0.5.4",
    "peft>=0.17.1",
    "torch>=2.8.0",
    "transformers>=4.56.1",
    "trl>=0.23.0",
    "dictionary-learning @ git+https://github.com/jkminder/dictionary_learning",
    "pyyaml>=6.0.2",
]
